requirements.txt.txt
opencv-python==4.8.0.74
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
pillow==10.0.0
pytesseract==0.3.10
numpy==1.25.2
